Separete into Price and GasPrice
