+++ src/components/SingleName/NameRegister/Price.js
@@ -1,13 +1,9 @@
-import React, { useState } from 'react'
+import React from 'react'
 import styled from '@emotion/styled/macro'
-import { useTranslation, Trans } from 'react-i18next'
+import { useTranslation } from 'react-i18next'
 import mq from 'mediaQuery'
 import EthVal from 'ethval'
 import { InlineLoader } from 'components/Loader'
-import DefaultInput from '../../Forms/Input'
-const GWEI = 1000000000
-const COMMIT_GAS_WEI = 42000
-const REGISTER_GAS_WEI = 240000
 
 const PriceContainer = styled('div')`
   width: 100%;
@@ -27,10 +23,6 @@ const Value = styled('div')`
   `}
 `
 
-const TotalValue = styled(Value)`
-  font-weight: 300;
-`
-
 const Description = styled('div')`
   font-family: Overpass;
   font-weight: 300;
@@ -48,12 +40,6 @@ const USD = styled('span')`
   `}
 `
 
-const Input = styled(DefaultInput)`
-  display: inline-block;
-  width: 4em;
-  margin-bottom: 0em;
-`
-
 const Price = ({
   loading,
   price,
@@ -64,10 +50,6 @@ const Price = ({
   underPremium
 }) => {
   const { t } = useTranslation()
-  const [gasPrice, setGasPrice] = useState(initialGasPrice)
-  const handleGasPrice = e => {
-    setGasPrice((e.target.value || 0) * GWEI)
-  }
 
   let ethPrice = <InlineLoader />
   let ethVal, basePrice, withPremium, usdPremium
@@ -85,69 +67,23 @@ const Price = ({
       usdPremium = ethVal.mul(ethUsdPrice).toFixed(2)
     }
   }
-  const commitGas = new EthVal(`${COMMIT_GAS_WEI * gasPrice}`).toEth()
-  const registerGas = new EthVal(`${REGISTER_GAS_WEI * gasPrice}`).toEth()
-  const gasPriceToGwei = new EthVal(`${gasPrice}`).toGwei()
-  const totalGas = commitGas.add(registerGas)
-  const totalGasInUsd = totalGas.mul(ethUsdPrice)
-  const buffer = ethVal.div(10)
-  const total = ethVal.add(buffer).add(totalGas)
-  const totalInUsd = total.mul(ethUsdPrice)
   return (
-    <>
-      <PriceContainer>
-        <Value>
-          {ethPrice} ETH
-          {ethVal && ethUsdPrice && (
-            <USD>
-              {withPremium}${usdPremium}
-              USD
-            </USD>
-          )}
-        </Value>
-        <Description>
-          {ethUsdPremiumPrice
-            ? t('pricer.pricePerAmount')
-            : t('pricer.registrationPriceLabel')}
-        </Description>
-      </PriceContainer>
-      <PriceContainer>
-        <Value>
-          {totalGas.toFixed(3)} ETH ({commitGas.toFixed(3)} ETH +{' '}
-          {registerGas.toFixed(3)} ETH){' '}
-          <Trans i18nKey="pricer.gasValue">
-            when the gas price is
-            <Input
-              value={gasPriceToGwei.toFixed(0)}
-              onChange={handleGasPrice}
-            />{' '}
-            Gwei
-          </Trans>
-          {ethVal && ethUsdPrice && (
-            <USD>
-              {' '}
-              = ${totalGasInUsd.toFixed(2)}
-              USD
-            </USD>
-          )}
-        </Value>
-        <Description>{t('pricer.gasDescription')}</Description>
-      </PriceContainer>
-      <PriceContainer>
-        <TotalValue>
-          {total.toFixed(3)} ETH ({ethVal.toFixed(3)} ETH + {buffer.toFixed(3)}{' '}
-          ETH + {totalGas.toFixed(3)} ETH)
-          {ethVal && ethUsdPrice && (
-            <USD>
-              {' '}
-              = ${totalInUsd.toFixed(2)}
-              USD
-            </USD>
-          )}
-        </TotalValue>
-        <Description>{t('pricer.totalDescription')}</Description>
-      </PriceContainer>
-    </>
+    <PriceContainer>
+      <Value>
+        {ethPrice} ETH
+        {ethVal && ethUsdPrice && (
+          <USD>
+            {withPremium}${usdPremium}
+            USD
+          </USD>
+        )}
+      </Value>
+      <Description>
+        {ethUsdPremiumPrice
+          ? t('pricer.pricePerAmount')
+          : t('pricer.registrationPriceLabel')}
+      </Description>
+    </PriceContainer>
   )
 }
